+++ lib/screens/scheduled_screen.dart
@@ -320,8 +320,6 @@ class _ScheduledScreenState extends State<ScheduledScreen> {
           notifiedDate: advance(bill.notifiedDate, i),
           notificationMethod: bill.notificationMethod,
           paymentMethod: bill.paymentMethod,
-          scheduledDate: advance(bill.scheduledDate, i),
-          confirmedPaidDate: advance(bill.confirmedPaidDate, i),
           note: bill.note,
           isTemplate: false,
           isAutoPaid: bill.isAutoPaid,